pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.1.0"
description = "My Python project"
authors = ["John Doe <<EMAIL>>"]
license = "MIT"

NIXPACKS_PYTHON_VERSION = "3.11"

[tool.poetry.dependencies]
python = "^3.11.4"
fastapi = "^0.108.0"
Telethon = "^1.33.1"
pydantic = "^1.10.0"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
uvicorn = "^0.25.0"
APScheduler = "^3.10.4"
SQLAlchemy = "^2.0.25"
asyncpg = "^0.29.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
